動画アップロードなどをプロフィールページに統合
